np float to float

+++ programs/analysis/2023/utilities.py
@@ -254,7 +254,7 @@ def fourier(data):
 # Function for extrapolating the limb darkening coefficient
 def get_u(temp_star, logg_star):
     # read in table with LD coefficients
-    table = np.loadtxt("tableu.txt", delimiter=' ', dtype={'names': ('logg', 'temp', 'Z', 'xi', 'u', 'Filt', 'Met'), 'formats': (np.float, np.float, np.float, np.float, np.float, '|S15', '|S15')}, usecols=(1,2,3,4,5,6,7))
+    table = np.loadtxt("tableu.txt", delimiter=' ', dtype={'names': ('logg', 'temp', 'Z', 'xi', 'u', 'Filt', 'Met'), 'formats': (float, float, float, float, float, '|S15', '|S15')}, usecols=(1,2,3,4,5,6,7))
     # shorten table to interesting values (z=0, xi=2, filter=B, method=L)
     tables = []
     for i in range(len(table)):
@@ -291,7 +291,7 @@ def get_u(temp_star, logg_star):
         f = scipy.interpolate.interp1d(logg, u,'cubic', bounds_error=False, fill_value='extrapolate')
         ynew = f(xnew)
         plt.plot(xnew, ynew, color='green')
-        x_want = round(np.float(f(logg_star)),2)
+        x_want = round(float(f(logg_star)),2)
         
     elif len(logg) == 0:
         plt.plot(logg1, u1, marker='o',linestyle='', color='green', label=str(temp1))
@@ -304,9 +304,9 @@ def get_u(temp_star, logg_star):
         plt.plot(xnew, y2new, color='blue')
         x1_want = f1(logg_star)
         x2_want = f2(logg_star)
-        print(round(np.float(x1_want),2))
-        print(round(np.float(x2_want),2))
-        x_want = round(np.float(np.mean([x1_want, x2_want])),2)
+        print(round(float(x1_want),2))
+        print(round(float(x2_want),2))
+        x_want = round(float(np.mean([x1_want, x2_want])),2)
     plt.legend()    
     return(x_want)
